+++ src/components/OrderBreakdown/LineItemBasePriceMaybe.js
@@ -6,19 +6,19 @@ import { LINE_ITEM_NIGHT, LINE_ITEM_DAY, propTypes } from '../../util/types';
 import css from './OrderBreakdown.module.css';
 
 const LineItemBasePriceMaybe = props => {
-  const { lineItems, unitType, intl } = props;
-  const isNightly = unitType === LINE_ITEM_NIGHT;
-  const isDaily = unitType === LINE_ITEM_DAY;
+  const { lineItems, code, intl } = props;
+  const isNightly = code === LINE_ITEM_NIGHT;
+  const isDaily = code === LINE_ITEM_DAY;
   const translationKey = isNightly
     ? 'OrderBreakdown.baseUnitNight'
     : isDaily
     ? 'OrderBreakdown.baseUnitDay'
     : 'OrderBreakdown.baseUnitQuantity';
 
-  // Find correct line-item for given unitType prop.
-  // It should be one of the following: 'line-item/night, 'line-item/day', 'line-item/units', or 'line-item/time'
+  // Find correct line-item for given code prop.
+  // It should be one of the following: 'line-item/night, 'line-item/day', 'line-item/hour', or 'line-item/item'
   // These are defined in '../../util/types';
-  const unitPurchase = lineItems.find(item => item.code === unitType && !item.reversal);
+  const unitPurchase = lineItems.find(item => item.code === code && !item.reversal);
 
   const quantity = unitPurchase ? unitPurchase.quantity.toString() : null;
   const unitPrice = unitPurchase ? formatMoney(intl, unitPurchase.unitPrice) : null;
@@ -36,7 +36,7 @@ const LineItemBasePriceMaybe = props => {
 
 LineItemBasePriceMaybe.propTypes = {
   lineItems: propTypes.lineItems.isRequired,
-  unitType: propTypes.lineItemUnitType.isRequired,
+  code: propTypes.lineItemUnitType.isRequired,
   intl: intlShape.isRequired,
 };